fix: typecast

+++ src/syncerd/bitcoin_syncer.rs
@@ -229,7 +229,7 @@ impl ElectrumRpc {
                                     break;
                                 }
                                 warn!("{:?} should be found in the history if we successfully queried `transaction_get` - reattempting {:?}", &tx_id, attempts);
-                                std::thread::sleep(Duration::from_secs_f32(0.5 * attempts));
+                                std::thread::sleep(Duration::from_secs_f32(0.5 * attempts as f32));
                             }
                             Err(err) => {
                                 trace!(
